backend/requirements.txt
fastapi
uvicorn
python-jose
python-dotenv
hvac
